Move visibleFileStatesSelector to own folder #3747
