+++ visualization/app/codeCharta/state/effects/addBlacklistItemsIfNotResultsInEmptyMap/addBlacklistItemsIfNotResultsInEmptyMap.effect.ts
@@ -3,7 +3,7 @@ import { MatDialog } from "@angular/material/dialog"
 import { Actions, createEffect, ofType } from "@ngrx/effects"
 import { filter, map, share, tap, withLatestFrom } from "rxjs"
 import { addBlacklistItems, addBlacklistItemsIfNotResultsInEmptyMap } from "../../store/fileSettings/blacklist/blacklist.actions"
-import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
+import { visibleFileStatesSelector } from "../../selectors/visibleFileStates/visibleFileStates.selector"
 import { blacklistSelector } from "../../store/fileSettings/blacklist/blacklist.selector"
 import { resultsInEmptyMap } from "./resultsInEmptyMap"
 import { ErrorDialogComponent } from "../../../ui/dialogs/errorDialog/errorDialog.component"
