+++ visualization/app/codeCharta/state/effects/autoFitCodeMapChange/autoFitCodeMap.effect.ts
@@ -4,7 +4,7 @@ import { Actions, createEffect, ofType } from "@ngrx/effects"
 import { switchMap, filter, skip, take, tap, combineLatest, withLatestFrom, first } from "rxjs"
 import { CcState } from "../../../codeCharta.model"
 import { ThreeMapControlsService } from "../../../ui/codeMap/threeViewer/threeMapControls.service"
-import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
+import { visibleFileStatesSelector } from "../../selectors/visibleFileStates/visibleFileStates.selector"
 import { layoutAlgorithmSelector } from "../../store/appSettings/layoutAlgorithm/layoutAlgorithm.selector"
 import { resetCameraIfNewFileIsLoadedSelector } from "../../store/appSettings/resetCameraIfNewFileIsLoaded/resetCameraIfNewFileIsLoaded.selector"
 import { RenderCodeMapEffect } from "../renderCodeMapEffect/renderCodeMap.effect"
