+++ visualization/app/codeCharta/state/effects/updateFileSettings/updateFileSettings.effect.ts
@@ -5,7 +5,7 @@ import { setState } from "../../store/state.actions"
 import { CcState } from "../../../codeCharta.model"
 import { map } from "rxjs"
 import { getVisibleFiles, isPartialState } from "../../../model/files/files.helper"
-import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
+import { visibleFileStatesSelector } from "../../selectors/visibleFileStates/visibleFileStates.selector"
 import { getMergedEdges } from "./utils/edges.merger"
 import { getMergedMarkedPackages } from "./utils/markedPackages.merger"
 import { getMergedBlacklist } from "./utils/blacklist.merger"
