+++ visualization/app/codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect.spec.ts
@@ -5,7 +5,7 @@ import { TestBed } from "@angular/core/testing"
 import { EffectsModule } from "@ngrx/effects"
 import { State, StoreModule } from "@ngrx/store"
 import { appReducers, setStateMiddleware } from "../../store/state.manager"
-import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
+import { visibleFileStatesSelector } from "../../selectors/visibleFileStates/visibleFileStates.selector"
 import { codeMapNodesSelector } from "../../selectors/accumulatedData/codeMapNodes.selector"
 import { getLastAction } from "../../../util/testUtils/store.utils"
 import { setAmountOfTopLabels } from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
