+++ visualization/app/codeCharta/state/selectors/accumulatedData/accumulatedData.selector.ts
@@ -8,7 +8,7 @@ import { getDeltaFile } from "./utils/getDeltaFile"
 import { addEdgeMetricsForLeaves } from "./utils/addEdgeMetricsForLeaves"
 import { blacklistSelector } from "../../store/fileSettings/blacklist/blacklist.selector"
 import { attributeTypesSelector } from "../../store/fileSettings/attributeTypes/attributeTypes.selector"
-import { visibleFileStatesSelector } from "../visibleFileStates.selector"
+import { visibleFileStatesSelector } from "../visibleFileStates/visibleFileStates.selector"
 import { metricDataSelector } from "./metricData/metricData.selector"
 import { createSelector } from "@ngrx/store"
 import { clone } from "../../../util/clone"
