+++ visualization/app/codeCharta/state/selectors/visibleFileStates/visibleFileStates.selector.ts
@@ -1,5 +1,5 @@
 import { createSelector } from "@ngrx/store"
-import { getVisibleFileStates } from "../../model/files/files.helper"
-import { filesSelector } from "../store/files/files.selector"
+import { getVisibleFileStates } from "../../../model/files/files.helper"
+import { filesSelector } from "../../store/files/files.selector"
 
 export const visibleFileStatesSelector = createSelector(filesSelector, getVisibleFileStates)
